refactor(map): clean all the classes for the debugging
