--- app/src/main/java/com/android/streetworkapp/SimpleData.kt
@@ -1,12 +1,0 @@
-package com.android.streetworkapp
-
-import kotlin.math.sqrt
-
-data class Point(val x: Double, val y: Double) {
-
-  fun distanceTo(p: Point): Double {
-    val dx = x - p.x
-    val dy = y - p.y
-    return sqrt(dx * dx + dy * dy)
-  }
-}
